requirements.txt
psycopg2-binary==2.8.3
Radar
postgresql app
